moved shouldFlyTo to map component
